Complete patientor exercise step11 feat: Extend the entry form to support all the entry types.
Note: A fake/hack type is created for submitting entry field due to an
error preventing the initialisation of fields that are not in the
selected type. The error could be be fixed by creating 3 seperate
components that only initialize the required fields when the type is
selected which I've chosen to avoid doing.